+++ src/jsonrpccpp/server/connectors/linuxtcpsocketserver.cpp
@@ -214,7 +214,7 @@ bool LinuxTcpSocketServer::WriteToSocket(const int& fd, const string& toWrite)
 			errorOccured = true;
 			CleanClose(fd);
 		}
-		else if(byteWritten < toSend.size())
+		else if(static_cast<size_t>(byteWritten) < toSend.size())
 		{
 			int len = toSend.size() - byteWritten;
 			toSend = toSend.substr(byteWritten + sizeof(char), len);
